nit: move some more files around
