+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/MongoDb/BsonSerialization/MongoBsonWriter.cs
@@ -5,9 +5,10 @@
 
 using System;
 using Datadog.Trace.DuckTyping;
+
 #pragma warning disable CS1591
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.MongoDb;
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.MongoDb.BsonSerialization;
 
 /// <summary>
 /// Duck Typing interface proxy for: https://github.com/mongodb/mongo-csharp-driver/blob/v2.8.x/src/MongoDB.Bson/IO/IBsonWriter.cs
